requirements.txt
# API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Notion API
notion-client>=2.0.0

# Google Workspace APIs
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0

# AI model dependencies
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0

# Testing
pytest>=7.0.0

# Code formatting and linting
black>=23.0.0
flake8>=6.0.0